index page code split
