+++ src/webportal/src/app/home/index.ejs
@@ -9,7 +9,6 @@
 </head>
 <body>
   <div id="content"></div>
-  <script>window.PAI_VERSION = <%=JSON.stringify(htmlWebpackPlugin.options.version)%></script>
   <script src="/scripts/env.js"></script>
 </body>
 </html>
